make mod split_sequence for test
